+++ R/00_namespace.R
@@ -2,7 +2,7 @@
 #' @importFrom yaml yaml.load_file
 #' @importFrom httr GET HEAD content timeout
 #' @importFrom rvest read_html html_nodes html_text
-#' @importFrom stringr str_squish
+#' @importFrom stringr str_squish regex str_detect
 #' @importFrom tidyRSS tidyfeed
 #' @importFrom stats setNames
 #' @importFrom mongolite mongo