+++ NEWS.md
@@ -1,3 +1,13 @@
+
+# rssData 25.07.0
+
+## New Features
+-  **filter feature**: It allows that RSS texts can be filtered based on keywords specified in the config file.
+- The filter location (**title** or **text**) is configurable via the config as well.
+
+## Updates
+- Added **two new sources** in the config file.
+
 # rssData 0.2.1
 
 ## Updates